0.18 add more requests features

+++ service/src/main/java/ru/ewm/service/event/dto/EventFullDto.java
@@ -21,7 +21,7 @@ public class EventFullDto {
 
     private String annotation;
     private CategoryDto category;
-    private Long confirmedRequests;
+    private int confirmedRequests;
     @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = DATE_TIME_PATTERN)
     private LocalDateTime createdOn;
     private String description;
